Adds internal package and move everything inside but the distribution package
