+++ pkg/internal/gguf/create.go
@@ -5,11 +5,10 @@ import (
 	"strings"
 	"time"
 
+	"github.com/docker/model-distribution/pkg/internal/partial"
+	"github.com/docker/model-distribution/pkg/internal/types"
 	v1 "github.com/google/go-containerregistry/pkg/v1"
 	gguf_parser "github.com/gpustack/gguf-parser-go"
-
-	"github.com/docker/model-distribution/pkg/partial"
-	"github.com/docker/model-distribution/pkg/types"
 )
 
 func NewModel(path string) (*Model, error) {
