+++ pkg/internal/store/errors.go
