+++ pkg/internal/store/testdata/dummy.gguf
