+++ pkg/internal/store/testdata/license.txt
