+++ pkg/internal/types/model.go
